+++ src/main/java/com/hfut/studentmanager/pojo/ClazzCourseTeacher.java
@@ -3,49 +3,49 @@ package com.hfut.studentmanager.pojo;
 
 public class ClazzCourseTeacher {
 
-  private long id;
-  private long clazzId;
-  private long gradeId;
-  private long courseId;
-  private long teacherId;
+  private Integer id;
+  private Integer clazzId;
+  private Integer gradeId;
+  private Integer courseId;
+  private Integer teacherId;
 
-  public long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
 
-  public long getClazzId() {
+  public Integer getClazzId() {
     return clazzId;
   }
 
-  public void setClazzId(long clazzId) {
+  public void setClazzId(Integer clazzId) {
     this.clazzId = clazzId;
   }
 
-  public long getGradeId() {
+  public Integer getGradeId() {
     return gradeId;
   }
 
-  public void setGradeId(long gradeId) {
+  public void setGradeId(Integer gradeId) {
     this.gradeId = gradeId;
   }
 
-  public long getCourseId() {
+  public Integer getCourseId() {
     return courseId;
   }
 
-  public void setCourseId(long courseId) {
+  public void setCourseId(Integer courseId) {
     this.courseId = courseId;
   }
 
-  public long getTeacherId() {
+  public Integer getTeacherId() {
     return teacherId;
   }
 
-  public void setTeacherId(long teacherId) {
+  public void setTeacherId(Integer teacherId) {
     this.teacherId = teacherId;
   }
 }
